% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_prop.R
\name{count_prop}
\alias{count_prop}
\title{Count and compute proportion in data frame column}
\usage{
count_prop(df, var, sort = FALSE)
}
\arguments{
\item{df}{data frame}

\item{var}{column in data frame}

\item{sort}{sort count}
}
\value{
A \link[tibble:tibble-package]{tibble} with the count and proportion
of the column
}
\description{
Count and compute proportion in data frame column
}
\examples{
df <- tibble::tibble(
   a = rnorm(5),
   b = rnorm(5),
   c = rnorm(5),
   d = rnorm(5))

df |> count_prop(a)
}
