% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_prop.R
\name{count_prop_mult}
\alias{count_prop_mult}
\title{Count and compute proportion in data frame column}
\usage{
count_prop_mult(df, rows, cols, sort = FALSE)
}
\arguments{
\item{df}{data frame}

\item{rows}{rows in data frame}

\item{cols}{cols in data frame}

\item{sort}{sort count}
}
\value{
A \link[tibble:tibble-package]{tibble} with the count and proportion
of the column
}
\description{
Count and compute proportion in data frame column
}
\examples{
options(scipen = 999)
ggplot2::diamonds |>
count_prop_mult(c(clarity, color), cut, sort = TRUE) |> head()
}
