% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurt.R
\name{kurt}
\alias{kurt}
\title{Compute kurtosis}
\usage{
kurt(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector}

\item{na.rm}{remove NAs}
}
\value{
vector
}
\description{
Compute kurtosis
}
\examples{
kurt(cars$speed)

kurt(cars$dist)
}
