% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentages.R
\name{pct_change}
\alias{pct_change}
\title{Calculate the percentage change between two values}
\usage{
pct_change(x, y)
}
\arguments{
\item{x}{newer value}

\item{y}{older value}
}
\value{
vector
}
\description{
The change in value is divided by the original value and then
multiplied by 100. If the result is positive, then it is an increase.
If the result is negative, then it is a decrease.
}
\examples{
pct_change(265, 471)
pct_change(471, 265)
}
