% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_diff.R
\name{pct_diff}
\alias{pct_diff}
\title{Calculate the percent difference between two values}
\usage{
pct_diff(x, y)
}
\arguments{
\item{x, y}{values to determine the percent difference between.}
}
\value{
vector
}
\description{
Percent difference is obtained by dividing the absolute value
of change by the average of the values and then multiplying it with 100.
Since the absolute value is taken for the change (or difference) in
values, the order of the numbers does not matter.
}
\examples{
pct_diff(265, 471)

pct_diff(471, 265)
}
