% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtskew.R
\name{rescale01}
\alias{rescale01}
\title{Rescales vector to a range from 0 to 1}
\usage{
rescale01(x)
}
\arguments{
\item{x}{vector}
}
\value{
vector with rescaled values
}
\description{
Rescales vector to a range from 0 to 1
}
\examples{
rescale01(c(-10, 0, 10))
rescale01(c(1:10, Inf, NA))
df <- tibble::tibble(
a = rnorm(5),
b = rnorm(5),
c = rnorm(5),
d = rnorm(5))
df |> dplyr::mutate(
a = rescale01(a),
b = rescale01(b),
c = rescale01(c),
d = rescale01(d))
df |> dplyr::mutate(dplyr::across(a:d, rescale01))
}
