% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtskew.R
\name{skew}
\alias{skew}
\title{Compute skewness}
\usage{
skew(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector}

\item{na.rm}{remove NAs}
}
\value{
vector
}
\description{
Compute skewness
}
\examples{
skew(cars$speed)
skew(cars$dist)
}
