% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summit.R
\name{summit}
\alias{summit}
\title{Summarizes var column in data}
\usage{
summit(data, var)
}
\arguments{
\item{data}{df}

\item{var}{col in df}
}
\value{
A \link[tibble:tibble-package]{tibble} with the summarized data
}
\description{
Summarizes var column in data
}
\examples{
df <- tibble::tibble(
   a = rnorm(5),
   b = rnorm(5),
   c = rnorm(5),
   d = rnorm(5),
   e = 1:5)

df |>
  dplyr::group_by(e) |>
  summit(a)
}
